Added Level shifter
Fixed hardcoded library but forking cusf kicad

Fixed UART misconnection

Added the CD74HC4050 to be used as an buffer for the SPI and UART connections. The document I used also uses a 3v3 voltage regulator but considering we are powering from our own 3v3 supply we should be fine.

I have added the 74HC4050 as a lib folder separately. Once it's merged into cusf-kicad I will delete this file.